+++ compatibility_spark-2/src/main/scala/za/co/absa/hyperdrive/compatibility/impl/SparkUtil.scala
@@ -18,12 +18,12 @@ package za.co.absa.hyperdrive.compatibility.impl
 import org.apache.hadoop.fs.Path
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.execution.streaming.{FileStreamSink, MetadataLogFileIndex}
-import za.co.absa.hyperdrive.compatibility.api.SparkCompatibilityProvider
+import za.co.absa.hyperdrive.compatibility.api.CompatibleSparkUtil
 
-object SparkCompatibilityUtil extends SparkCompatibilityProvider {
+object SparkUtil extends CompatibleSparkUtil {
   override def createMetadataLogFileIndex(spark: SparkSession, destination: String): MetadataLogFileIndex =
-    new MetadataLogFileIndex(spark, new Path(destination), Map.empty, None)
+    new MetadataLogFileIndex(spark, new Path(destination), None)
 
   override def hasMetadata(spark: SparkSession, destination: String): Boolean =
-    FileStreamSink.hasMetadata(Seq(destination), spark.sparkContext.hadoopConfiguration, spark.sessionState.conf)
+    FileStreamSink.hasMetadata(Seq(destination), spark.sparkContext.hadoopConfiguration)
 }
