+++ ingestor-default/src/main/scala/org/apache/spark/sql/kafka010/KafkaSourceOffsetProxy.scala
@@ -17,10 +17,10 @@ package org.apache.spark.sql.kafka010
 
 import org.apache.kafka.common.TopicPartition
 import org.apache.spark.sql.execution.streaming.Offset
-import za.co.absa.hyperdrive.compatibility.provider.CompatibleOffset
+import za.co.absa.hyperdrive.compatibility.provider.CompatibleOffsetProvider
 
 object KafkaSourceOffsetProxy {
-  def getPartitionOffsets(offset: CompatibleOffset.Type): Map[TopicPartition, Long] = {
+  def getPartitionOffsets(offset: CompatibleOffsetProvider.Type): Map[TopicPartition, Long] = {
     // TODO: Remove casting as soon as spark kafka is migrated to org.apache.spark.sql.connector.read.streaming.Offset
     KafkaSourceOffset.getPartitionOffsets(offset.asInstanceOf[Offset])
   }
